pyproject.toml
[project]
name = "automation-auditor"
version = "0.1.0"
description = "LangGraph Digital Courtroom for autonomous code/report auditing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "docling>=2.0.0",
]

[project.optional-dependencies]
vision = [
    "google-generativeai>=0.8.0",
]

[project.scripts]
run-audit = "src.run_audit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
